+++ app/Filament/Resources/CarSearchResource/Pages/ListCarSearches.php
@@ -3,9 +3,17 @@
 namespace App\Filament\Resources\CarSearchResource\Pages;
 
 use App\Filament\Resources\CarSearchResource;
+use Filament\Actions;
 use Filament\Resources\Pages\ListRecords;
 
 class ListCarSearches extends ListRecords
 {
     protected static string $resource = CarSearchResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\CreateAction::make(),
+        ];
+    }
 }